feat: Implement Uplink 'bucket'
Implement the Uplink 'bucket' and add a new variant to the Error type
for representing internal errors because it's something that could
happen when creating a 'bucket' from the Uplink c-bindings 'bucket' type
and it's likely that will be other cases where an internal can happen.

Also, updates the README, marking the task for implementing 'bucket' as
done.