refactor: Create alias for Result with specific Error
Create an alias type for the Result type that the Err case is always of
the Error type published by this crate.

All the public functions and method of this crate that may return an
error, with the exception of the ones that fulfill the implementation of
external traits, return an error of the Error type published by this
crate.

The alias makes those functions and methods signatures less repetitive
and more concise and follow the patterns used in the standard library,
as std::io, and the official Rust book
(https://doc.rust-lang.org/book/ch19-04-advanced-types.html?highlight=alias#creating-type-synonyms-with-type-aliases).