fix: Don't expose outside crate the whole error module
The error module was public outside of the crate and it should only
expose the Error enum and the types which are used by its variants and
the fields or methods which provide information about the error but
nothing else, like the constructors to create Errors.

While publishing the constructors or other things of the error module
may be good for using them by consumers, the goal of this crate isn't
tor provide an Error type for being used outside, hence, the more that
it publish outside the easier will be to introduce changes which break
backward compatibility.