refactor: Publish access module & rename Access struct
I've realized that it looks much better to expose all the types related
to an access grant under the module access so publish the entire
`access` module.

For not stuttering, rename the `Access` struct to `Grant` because it
represents an access grant and now it's exposed outside of the crate
under the module `access` rather than directly expose it.